ucm2: Qualcomm: x1e80100: T14s: add Microphone Mute LED
Add support for the Microphone Mute LED in the Keyboard's MicMute Key,
which is exposed by the lenovo-thinkpad-t14s EC driver that landed in
the 6.18 kernel and uses the "audio-micmute" trigger by default. This
works by creating a virtual capture switch and binding the LED to it.

Something similar should be done for the speaker mute LED, but I'm
not sure how that is supposed to look like. The Speaker Mute LED
automatically toggles when the Speaker Mute Key is pressed, so it
seems to work without explicit handling. But it can get out-of-sync,
if the mute status is modified in software without using the key.

Signed-off-by: Sebastian Reichel <sre@kernel.org>